FortiBot-NOC-CLI/requirements.txt
click>=8.0
rich>=13.0
anthropic>=0.40.0
paramiko>=3.0
requests>=2.28
urllib3>=1.26
pyyaml>=6.0
cryptography>=41.0
